Fixed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@10732 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/challenges/challenge_data.cpp
@@ -159,7 +159,7 @@ ChallengeData::ChallengeData(const std::string& filename)
         }
         
         int time = -1;
-        if (requirements_node->get("time", &time)) m_time[d] = time;
+        if (requirements_node->get("time", &time)) m_time[d] = (float)time;
         
         if (m_time[d] < 0 && m_position[d] < 0) error("position/time");
